+++ packages/react/src/QuestionnaireForm/QuestionnaireFormItem/QuestionnaireFormItem.tsx
@@ -32,63 +32,71 @@ import {
 export interface QuestionnaireFormItemProps {
   item: QuestionnaireItem;
   index: number;
-  allResponses: QuestionnaireResponseItem[];
-  currentResponseItems?: QuestionnaireResponseItem[];
-  groupSequence?: number;
+  response?: QuestionnaireResponseItem;
   onChange: (newResponseItem: QuestionnaireResponseItem) => void;
 }
 
 export function QuestionnaireFormItem(props: QuestionnaireFormItemProps): JSX.Element | null {
   const item = props.item;
-  const index = props.index;
+  const response = props.response;
+
+  if (!response) {
+    return null;
+  }
 
   function onChangeAnswer(
-    newResponseAnswer: QuestionnaireResponseItemAnswer | QuestionnaireResponseItemAnswer[],
-    repeatedIndex?: number
+    newResponseAnswer: QuestionnaireResponseItemAnswer | QuestionnaireResponseItemAnswer[]
   ): void {
-    const number = repeatedIndex ?? 0;
-    const responses = props.currentResponseItems?.filter((r) => r.linkId === item.linkId) ?? [];
-
     let updatedAnswers: QuestionnaireResponseItemAnswer[];
     if (Array.isArray(newResponseAnswer)) {
       // It's a multi-select case, so use the array directly.
       updatedAnswers = newResponseAnswer;
+    } else if (props.index !== 0) {
+      if (props.index >= (props.response?.answer?.length ?? 0)) {
+        updatedAnswers = (props.response?.answer ?? []).concat([newResponseAnswer]);
+      } else {
+        const newAnswers = (props.response?.answer ?? []).map((a, idx) =>
+          idx === props.index ? newResponseAnswer : a
+        ) as QuestionnaireResponseItemAnswer[];
+        updatedAnswers = newAnswers ?? [];
+      }
     } else {
-      // It's a single answer case.
-      updatedAnswers = updateAnswerArray(responses[0]?.answer ?? [], number, newResponseAnswer);
+      updatedAnswers = [newResponseAnswer];
     }
     props.onChange({
-      id: responses[0].id,
-      linkId: item.linkId,
+      id: response?.id,
+      linkId: response?.linkId,
       text: item.text,
       answer: updatedAnswers,
     });
   }
 
-  const type = item.type as QuestionnaireItemType;
+  const type = item.type;
   if (!type) {
     return null;
   }
 
-  const name = item.linkId;
+  const name = response.id;
   if (!name) {
     return null;
   }
 
   const initial = item.initial && item.initial.length > 0 ? item.initial[0] : undefined;
   const defaultValue =
-    getCurrentAnswer(props.allResponses, item, props.index, props.groupSequence) ??
+    getCurrentAnswer(response, props.index) ??
     getTypedPropertyValue({ type: 'QuestionnaireItemInitial', value: initial }, 'value');
 
   switch (type) {
+    case QuestionnaireItemType.display:
+      return <p key={props.item.linkId}>{props.item.text}</p>;
     case QuestionnaireItemType.boolean:
       return (
         <CheckboxFormSection key={props.item.linkId} title={props.item.text} htmlFor={props.item.linkId}>
           <Checkbox
             id={props.item.linkId}
-            name={props.item.linkId}
+            name={name}
             defaultChecked={defaultValue?.value}
-            onChange={(e) => onChangeAnswer({ valueBoolean: e.currentTarget.checked }, index)}
+            onChange={(e) => onChangeAnswer({ valueBoolean: e.currentTarget.checked })}
           />
         </CheckboxFormSection>
       );
@@ -101,7 +109,7 @@ export function QuestionnaireFormItem(props: QuestionnaireFormItemProps): JSX.El
           name={name}
           required={item.required}
           defaultValue={defaultValue?.value}
-          onChange={(e) => onChangeAnswer({ valueDecimal: e.currentTarget.valueAsNumber }, index)}
+          onChange={(e) => onChangeAnswer({ valueDecimal: e.currentTarget.valueAsNumber })}
         />
       );
     case QuestionnaireItemType.integer:
@@ -113,7 +121,7 @@ export function QuestionnaireFormItem(props: QuestionnaireFormItemProps): JSX.El
           name={name}
           required={item.required}
           defaultValue={defaultValue?.value}
-          onChange={(e) => onChangeAnswer({ valueInteger: e.currentTarget.valueAsNumber }, index)}
+          onChange={(e) => onChangeAnswer({ valueInteger: e.currentTarget.valueAsNumber })}
         />
       );
     case QuestionnaireItemType.date:
@@ -124,7 +132,7 @@ export function QuestionnaireFormItem(props: QuestionnaireFormItemProps): JSX.El
           name={name}
           required={item.required}
           defaultValue={defaultValue?.value}
-          onChange={(e) => onChangeAnswer({ valueDate: e.currentTarget.value }, index)}
+          onChange={(e) => onChangeAnswer({ valueDate: e.currentTarget.value })}
         />
       );
     case QuestionnaireItemType.dateTime:
@@ -133,7 +141,7 @@ export function QuestionnaireFormItem(props: QuestionnaireFormItemProps): JSX.El
           name={name}
           required={item.required}
           defaultValue={defaultValue?.value}
-          onChange={(newValue: string) => onChangeAnswer({ valueDateTime: newValue }, index)}
+          onChange={(newValue: string) => onChangeAnswer({ valueDateTime: newValue })}
         />
       );
     case QuestionnaireItemType.time:
@@ -144,7 +152,7 @@ export function QuestionnaireFormItem(props: QuestionnaireFormItemProps): JSX.El
           name={name}
           required={item.required}
           defaultValue={defaultValue?.value}
-          onChange={(e) => onChangeAnswer({ valueTime: e.currentTarget.value }, index)}
+          onChange={(e) => onChangeAnswer({ valueTime: e.currentTarget.value })}
         />
       );
     case QuestionnaireItemType.string:
@@ -155,7 +163,7 @@ export function QuestionnaireFormItem(props: QuestionnaireFormItemProps): JSX.El
           name={name}
           required={item.required}
           defaultValue={defaultValue?.value}
-          onChange={(e) => onChangeAnswer({ valueString: e.currentTarget.value }, index)}
+          onChange={(e) => onChangeAnswer({ valueString: e.currentTarget.value })}
         />
       );
     case QuestionnaireItemType.text:
@@ -165,7 +173,7 @@ export function QuestionnaireFormItem(props: QuestionnaireFormItemProps): JSX.El
           name={name}
           required={item.required}
           defaultValue={defaultValue?.value}
-          onChange={(e) => onChangeAnswer({ valueString: e.currentTarget.value }, index)}
+          onChange={(e) => onChangeAnswer({ valueString: e.currentTarget.value })}
         />
       );
     case QuestionnaireItemType.attachment:
@@ -174,7 +182,7 @@ export function QuestionnaireFormItem(props: QuestionnaireFormItemProps): JSX.El
           <AttachmentInput
             name={name}
             defaultValue={defaultValue?.value}
-            onChange={(newValue) => onChangeAnswer({ valueAttachment: newValue }, index)}
+            onChange={(newValue) => onChangeAnswer({ valueAttachment: newValue })}
           />
         </Group>
       );
@@ -185,7 +193,7 @@ export function QuestionnaireFormItem(props: QuestionnaireFormItemProps): JSX.El
           required={item.required}
           targetTypes={getQuestionnaireItemReferenceTargetTypes(item)}
           defaultValue={defaultValue?.value}
-          onChange={(newValue) => onChangeAnswer({ valueReference: newValue }, index)}
+          onChange={(newValue) => onChangeAnswer({ valueReference: newValue })}
         />
       );
     case QuestionnaireItemType.quantity:
@@ -194,7 +202,7 @@ export function QuestionnaireFormItem(props: QuestionnaireFormItemProps): JSX.El
           name={name}
           required={item.required}
           defaultValue={defaultValue?.value}
-          onChange={(newValue) => onChangeAnswer({ valueQuantity: newValue }, index)}
+          onChange={(newValue) => onChangeAnswer({ valueQuantity: newValue })}
           disableWheel
         />
       );
@@ -206,10 +214,8 @@ export function QuestionnaireFormItem(props: QuestionnaireFormItemProps): JSX.El
             name={name}
             item={item}
             initial={initial}
-            allResponses={props.allResponses}
-            index={index}
-            groupSequence={props.groupSequence}
-            onChangeAnswer={(e) => onChangeAnswer(e, index)}
+            response={response}
+            onChangeAnswer={(e) => onChangeAnswer(e)}
           />
         );
       } else {
@@ -218,8 +224,8 @@ export function QuestionnaireFormItem(props: QuestionnaireFormItemProps): JSX.El
             name={name}
             item={item}
             initial={initial}
-            allResponses={props.allResponses}
-            onChangeAnswer={(e) => onChangeAnswer(e, index)}
+            response={response}
+            onChangeAnswer={(e) => onChangeAnswer(e)}
           />
         );
       }
@@ -232,14 +238,12 @@ interface QuestionnaireChoiceInputProps {
   name: string;
   item: QuestionnaireItem;
   initial: QuestionnaireItemInitial | undefined;
-  allResponses: QuestionnaireResponseItem[];
-  index?: number;
-  groupSequence?: number;
+  response: QuestionnaireResponseItem;
   onChangeAnswer: (newResponseAnswer: QuestionnaireResponseItemAnswer | QuestionnaireResponseItemAnswer[]) => void;
 }
 
 function QuestionnaireChoiceDropDownInput(props: QuestionnaireChoiceInputProps): JSX.Element {
-  const { name, item, initial } = props;
+  const { name, item, initial, response } = props;
   const initialValue = getTypedPropertyValue({ type: 'QuestionnaireItemInitial', value: initial }, 'value') as
     | TypedValue
     | undefined;
@@ -256,12 +260,11 @@ function QuestionnaireChoiceDropDownInput(props: QuestionnaireChoiceInputProps):
   }
 
   const defaultValue =
-    getCurrentAnswer(props.allResponses, item, props.index, props.groupSequence) ??
-    getTypedPropertyValue({ type: 'QuestionnaireItemInitial', value: initial }, 'value');
+    getCurrentAnswer(response) ?? getTypedPropertyValue({ type: 'QuestionnaireItemInitial', value: initial }, 'value');
 
   if (item.repeats) {
     const { propertyName, data } = formatSelectData(props.item);
-    const currentAnswer = getCurrentMultiSelectAnswer(props.allResponses, item, props.groupSequence);
+    const currentAnswer = getCurrentMultiSelectAnswer(response);
 
     return (
       <MultiSelect
@@ -302,7 +305,7 @@ function QuestionnaireChoiceDropDownInput(props: QuestionnaireChoiceInputProps):
 }
 
 function QuestionnaireChoiceSetInput(props: QuestionnaireChoiceInputProps): JSX.Element {
-  const { name, item, initial, onChangeAnswer, allResponses } = props;
+  const { name, item, initial, onChangeAnswer, response } = props;
   if (item.answerValueSet) {
     return (
       <CodingInput
@@ -317,14 +320,14 @@ function QuestionnaireChoiceSetInput(props: QuestionnaireChoiceInputProps): JSX.
       name={name}
       item={item}
       initial={initial}
-      allResponses={allResponses}
+      response={response}
       onChangeAnswer={onChangeAnswer}
     />
   );
 }
 
 function QuestionnaireChoiceRadioInput(props: QuestionnaireChoiceInputProps): JSX.Element {
-  const { name, item, initial, onChangeAnswer } = props;
+  const { name, item, initial, onChangeAnswer, response } = props;
   const valueElementDefinition = getElementDefinition('QuestionnaireItemAnswerOption', 'value[x]');
   const initialValue = getTypedPropertyValue({ type: 'QuestionnaireItemInitial', value: initial }, 'value') as
     | TypedValue
@@ -348,7 +351,7 @@ function QuestionnaireChoiceRadioInput(props: QuestionnaireChoiceInputProps): JS
     }
   }
 
-  const defaultAnswer = getCurrentAnswer(props.allResponses, item, props.index, props.groupSequence);
+  const defaultAnswer = getCurrentAnswer(response);
   const answerLinkId = getCurrentRadioAnswer(options, defaultAnswer);
 
   return (
@@ -383,21 +386,40 @@ function QuestionnaireChoiceRadioInput(props: QuestionnaireChoiceInputProps): JS
   );
 }
 
-function updateAnswerArray(
-  answers: QuestionnaireResponseItemAnswer[],
-  index: number,
-  newResponseAnswer: QuestionnaireResponseItemAnswer
-): QuestionnaireResponseItemAnswer[] {
-  if (index < answers.length) {
-    answers[index] = newResponseAnswer;
-    return answers;
-  } else {
-    for (let i = answers.length; i < index; i++) {
-      answers.push({});
-    }
-    answers.push(newResponseAnswer);
-    return answers;
+function getItemValue(answer: QuestionnaireResponseItemAnswer): TypedValue {
+  const itemValue = getTypedPropertyValue({ type: 'QuestionnaireItemAnswer', value: answer }, 'value') as TypedValue;
+  return itemValue;
+}
+
+function getCurrentAnswer(response: QuestionnaireResponseItem, index: number = 0): TypedValue {
+  const results = response.answer;
+  return getItemValue(results?.[index] ?? {});
+}
+
+function getCurrentMultiSelectAnswer(response: QuestionnaireResponseItem): string[] {
+  const results = response.answer;
+  if (!results) {
+    return [];
   }
+  const typedValues = results.map((a) => getItemValue(a));
+  return typedValues.map((type) => formatCoding(type?.value) || type?.value);
+}
+
+function getCurrentRadioAnswer(options: [string, TypedValue][], defaultAnswer: TypedValue): string | undefined {
+  return options.find((option) => deepEquals(option[1].value, defaultAnswer?.value))?.[0];
+}
+
+function typedValueToString(typedValue: TypedValue | undefined): string | undefined {
+  if (!typedValue) {
+    return undefined;
+  }
+  if (typedValue.type === 'CodeableConcept') {
+    return typedValue.value.coding[0].display;
+  }
+  if (typedValue.type === 'Coding') {
+    return typedValue.value.display;
+  }
+  return typedValue.value.toString();
 }
 
 function isDropDownChoice(item: QuestionnaireItem): boolean {
@@ -438,68 +460,4 @@ function formatSelectData(item: QuestionnaireItem): FormattedData {
 
 function getValueAndLabel(option: QuestionnaireItemAnswerOption, propertyName: string): string | undefined {
   return formatCoding(option.valueCoding) || option[propertyName as keyof QuestionnaireItemAnswerOption]?.toString();
-}
-
-function typedValueToString(typedValue: TypedValue | undefined): string | undefined {
-  if (!typedValue) {
-    return undefined;
-  }
-  if (typedValue.type === 'CodeableConcept') {
-    return typedValue.value.coding[0].display;
-  }
-  if (typedValue.type === 'Coding') {
-    return typedValue.value.display;
-  }
-  return typedValue.value.toString();
-}
-
-function getItemsByLinkId(allResponses: QuestionnaireResponseItem[], linkId: string): QuestionnaireResponseItem[] {
-  let result: QuestionnaireResponseItem[] = [];
-
-  for (const item of allResponses) {
-    // If the linkId matches, add it to the result array
-    if (item.linkId === linkId) {
-      result.push(item);
-    }
-
-    // If the current item has nested items, search them too
-    if (item.item) {
-      result = result.concat(getItemsByLinkId(item.item, linkId));
-    }
-  }
-  return result;
-}
-
-function getItemValue(answer: QuestionnaireResponseItemAnswer): TypedValue {
-  const itemValue = getTypedPropertyValue({ type: 'QuestionnaireItemAnswer', value: answer }, 'value') as TypedValue;
-  return itemValue;
-}
-
-function getCurrentAnswer(
-  allResponses: QuestionnaireResponseItem[],
-  item: QuestionnaireItem,
-  index: number = 0,
-  groupSequence: number = 0
-): TypedValue {
-  const results = getItemsByLinkId(allResponses, item.linkId ?? '');
-  const selectedItem = results[groupSequence]?.answer;
-  return getItemValue(selectedItem?.[index] ?? {});
-}
-
-function getCurrentMultiSelectAnswer(
-  allResponses: QuestionnaireResponseItem[],
-  item: QuestionnaireItem,
-  groupSequence: number = 0
-): string[] {
-  const results = getItemsByLinkId(allResponses, item.linkId ?? '');
-  const selectedItem = results[groupSequence]?.answer;
-  if (!selectedItem) {
-    return [];
-  }
-  const typedValues = selectedItem.map((a) => getItemValue(a));
-  return typedValues.map((type) => formatCoding(type?.value) || type?.value);
-}
-
-function getCurrentRadioAnswer(options: [string, TypedValue][], defaultAnswer: TypedValue): string | undefined {
-  return options.find((option) => deepEquals(option[1].value, defaultAnswer?.value))?.[0];
 }
